% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{adjectives}
\alias{adjectives}
\title{Adjectives}
\format{
An object of class \code{character} of length 968.
}
\usage{
adjectives
}
\description{
A word bank of 968 adjectives
}
\keyword{datasets}
