% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomia_classes.R
\name{anom_generate_name}
\alias{anom_generate_name}
\title{Generate words from a template}
\usage{
anom_generate_name(template, n)
}
\arguments{
\item{template}{An object of class "anomia_template".}

\item{n}{The number of names to generate.}
}
\description{
Generate words from a template
}
