% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks}
\alias{name_chunks}
\title{Give a name to unnamed code blocks in an R Markdown document}
\usage{
name_chunks(path, template, unname_chunks = FALSE, overwrite)
}
\arguments{
\item{path}{The path to an R Markdown file}

\item{template}{An anomia name template class}

\item{unname_chunks}{Should existing chunk names be removed before generating new names.
This will try to preserve the "setup" block.}

\item{overwrite}{Should the modified R Markdown file be overwritten.}
}
\description{
Give a name to unnamed code blocks in an R Markdown document
}
