% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomia_classes.R
\name{new_combo}
\alias{new_combo}
\title{Constructor method for combination template}
\usage{
new_combo(
  pattern = list(),
  prefix = character(),
  suffix = character(),
  sep = character(),
  case = character()
)
}
\arguments{
\item{pattern}{A list of word banks to use for names.}

\item{prefix}{A string to prepend to the names. Can be left NULL.}

\item{suffix}{A string to append to the names. Can be left NULL.}

\item{sep}{A string used for separating the prefix, words, and suffix.}

\item{case}{The capitalization used for the words. Can be "lower" (default),
"upper", "title", or "sentence".}
}
\description{
Constructor method for combination template
}
