% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomia_classes.R
\name{validate_combo}
\alias{validate_combo}
\title{Validator method for combination template}
\usage{
validate_combo(x)
}
\arguments{
\item{x}{A combo template}
}
\description{
Validator method for combination template
}
